Remove dependency on `DecimalFormatter`
As the goal is to turn this into a multiplatform project, we can't really depend on any JDK specific dependencies.

The replacement function `formatCssDecimal` passes all the current tests, so it should be producing proper output, but I'm not 100% sure as it's something that I just threw together in a few minutes, without any proper understanding of how `DecimalFormatter` actually works.